+++ packages/codesandbox/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.5.20
+
+### Patch Changes
+
+- Updated dependencies [128edac]
+  - computesdk@1.18.0
+  - @computesdk/provider@1.0.14
+
 ## 1.5.19
 
 ### Patch Changes