+++ packages/e2b/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @computesdk/e2b
 
+## 1.7.20
+
+### Patch Changes
+
+- Updated dependencies [128edac]
+  - computesdk@1.18.0
+  - @computesdk/provider@1.0.14
+
 ## 1.7.19
 
 ### Patch Changes